Move payInvoiceList sync-schema to separate file
